+++ NEWS.md
@@ -8,6 +8,7 @@ Red Panda C++ Version 2.9
   - enhancement: Auto backup editing contents. (Save editing contents 3 seconds after input stopped. Auto delete when editor successfully closed)
   - enhancement: Add "Auto backup editing contents" option in options/editor/auto save
   - fix: rename project file will wrongly set it's encoding to 'ASCII';
+  - fix: Project's file encoding is wrongly set to 'AUTO' when load project.
 
 Red Panda C++ Version 2.8